ref: cleaner folder structure and some cleaning
