+++ Pages/Projects/Cards/ProjectsCard.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { Card, CardFooter } from "@nextui-org/react";
 import React from "react";
-import Image, { ImageLoaderProps } from "next/image";
+import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { motion } from "framer-motion";
