+++ public/sitemap-0.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>https://www.furkancengiz.software/projects/opengraph-image-wdsou1.jpg</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/about/opengraph-image-1itp17.jpg</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/icon.svg</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/opengraph-image.jpg</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/projects</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/projects/Learnification</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/projects/Turkish-Dictionary-using-TDK-API</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/projects/Meme-Generator</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/projects/Frontend-Mentor-Challenges</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
-<url><loc>https://www.furkancengiz.software/projects/Turkish-Dictionary</loc><lastmod>2023-11-07T06:12:53.603Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/opengraph-image.jpg</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/projects/opengraph-image-wdsou1.jpg</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/about/opengraph-image-1itp17.jpg</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/icon.svg</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/projects</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/projects/Learnification</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/projects/Turkish-Dictionary-using-TDK-API</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/projects/Meme-Generator</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/projects/Frontend-Mentor-Challenges</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software/projects/Turkish-Dictionary</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.furkancengiz.software</loc><lastmod>2023-11-07T14:37:13.803Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
 </urlset>
